refactor into smaller sections
